+++ JAG3D/src/org/applied_geodesy/jag3d/sql/SQLManager.java
@@ -67,9 +67,9 @@ import org.applied_geodesy.jag3d.ui.dialog.chart.SignType;
 import org.applied_geodesy.jag3d.ui.graphic.UIGraphicPaneBuilder;
 import org.applied_geodesy.jag3d.ui.graphic.sql.SQLGraphicManager;
 import org.applied_geodesy.jag3d.ui.io.reader.ImportOption;
-import org.applied_geodesy.jag3d.ui.io.report.FTLReport;
 import org.applied_geodesy.jag3d.ui.io.writer.ExportOption;
 import org.applied_geodesy.jag3d.ui.io.writer.ExportOption.ExportResultType;
+import org.applied_geodesy.jag3d.ui.io.writer.report.FTLReport;
 import org.applied_geodesy.jag3d.ui.metadata.MetaData;
 import org.applied_geodesy.jag3d.ui.metadata.UIMetaDataPaneBuilder;
 import org.applied_geodesy.jag3d.ui.propertiespane.UICongruenceAnalysisPropertiesPane;
