+++ JAG3D/src/org/applied_geodesy/jag3d/ui/io/reader/xml/AngularUnit.java
@@ -19,24 +19,8 @@
 *                                                                      *
 ***********************************************************************/
 
-package org.applied_geodesy.jag3d.ui.io.xml;
+package org.applied_geodesy.jag3d.ui.io.reader.xml;
 
-import org.applied_geodesy.util.io.xml.XMLNamespaceContext;
-import org.w3c.dom.Attr;
-import org.w3c.dom.Document;
-
-/**
- * @author Michael Loesler
- */
-class JobXMLNamespaceContext extends XMLNamespaceContext {
-	JobXMLNamespaceContext(Document document) {
-		super(document, "", "http://www.trimble.com/schema/JobXML/5_0");
-	}
-	
-	@Override
-	public String getPrefix(Attr attribute) {
-		if (attribute.getNodeValue().toLowerCase().contains("trimble.com/schema/jobxml"))
-			return "";
-		return this.getDefaultPrefix();
-	}
+enum AngularUnit {
+	RADIAN, GRADIAN, DEGREE, DDMMSSss;
 }
