+++ JAG3D/src/org/applied_geodesy/jag3d/ui/menu/UIMenuBuilder.java
@@ -71,10 +71,10 @@ import org.applied_geodesy.jag3d.ui.io.reader.ObservationFlatFileReader;
 import org.applied_geodesy.jag3d.ui.io.reader.PointFlatFileReader;
 import org.applied_geodesy.jag3d.ui.io.reader.VerticalDeflectionFlatFileReader;
 import org.applied_geodesy.jag3d.ui.io.reader.ZFileReader;
-import org.applied_geodesy.jag3d.ui.io.report.FTLReport;
+import org.applied_geodesy.jag3d.ui.io.reader.xml.HeXMLFileReader;
+import org.applied_geodesy.jag3d.ui.io.reader.xml.JobXMLFileReader;
 import org.applied_geodesy.jag3d.ui.io.sql.OADBReader;
-import org.applied_geodesy.jag3d.ui.io.xml.HeXMLFileReader;
-import org.applied_geodesy.jag3d.ui.io.xml.JobXMLFileReader;
+import org.applied_geodesy.jag3d.ui.io.writer.report.FTLReport;
 import org.applied_geodesy.jag3d.ui.tree.TreeItemValue;
 import org.applied_geodesy.jag3d.ui.tree.UITreeBuilder;
 import org.applied_geodesy.juniform.ui.JUniForm;
